requirements.txt
Flask
requests
redis
python-dotenv
alpha-vantage
pandas
pytest
pytest-mock
gunicorn
prometheus-flask-exporter
matplotlib
